feat(power): 储电池松散聚合成组, 容量与速率线性叠加
同档储电六向相邻自动并组, 点哪一块看到的都是整组读数, 但不引入固定
形状的多方块结构: 玩家怎么摆都算一个整体, 拆掉一块只是组变小。

分组索引 PowerCellGroupManager 完全复刻 EnergyNetworkManager 的增量
维护范式 —— 并组只找六向同档邻组, 拆组只对该组剩余成员重分连通分量,
读写路径零 flood-fill; 每 ServerLevel 一份并随 LevelEvent.Unload 释放。
分档判定走 spec 字段而不是 instanceof, 因为三档共用同一个
BlockEntityType, 用 instanceof 会并出三档混装组。

组本身不持有余额, 每块仍把自己那份写在自己的 NBT 里。容量与速率按成员
数线性叠加, 内部一律 long, 对外经 capability 暴露时饱和截断到
Integer.MAX_VALUE —— 三块未来储电 26.5 亿已越界, 强转会得到负数。

端点去重选"组速率在本 tick 内按已服务量扣减", 没选"整组只登记一个代表
端点": 代表端点方案要求那一块必须恰好贴着线缆, 玩家把 2x2 堆起来而线缆
只碰到右下角时整组会直接断电。共享额度让任意一面都能当接触点, 第一个被
服务到的面吃掉整份速率, 其余面本 tick 返回 0, 从而堵住"贴几个面就把速率
乘几遍"的刷电口子。额度按 gameTime 惰性清零, 与 GeneratorBlockEntity
的 extractionTick 同范式, 不依赖 BE ticker 与结算的先后顺序。

均摊两趟: 先按成员数等分, 装不下的再顺位溢流, 让"拆掉任意一块损失的份额"
接近平均值而不是取决于玩家拆哪一块。成员上限 64。

GUI 数据槽由 5 个 int 改 5 个 long: 原版同步包按 16 位写值, 而 64 块未来
储电的组容量是 566 亿 FE, 连 int 都装不下, 故每个读数切四段传输。

审查抓出的两条同源缺陷一并修复:
- 类注释宣称"拆掉一块它就带走自己那份", 实测三张战利品表都只有一条纯
  item 条目, 破坏储电等于余额湮灭。GUI 换成整组读数后这条销毁路径还被
  藏进了"一个大电池"的假象背后。补 copy_nbt 让注释成真, 并让余额为 0 时
  不写 storedFe 键, 否则挖下来的空储电会带 BlockEntityTag 而无法与合成品
  堆叠。
- group() 在索引查不到本坐标时抛 IllegalStateException, 而它挂在每 tick
  的电网结算路径上, "BE 已可解析但尚未进索引"是 Forge 真实存在的窗口。
  改为可空 + 每 BE 只 WARN 一次; 读数降级成单块组, 但收发一律返回 0 ——
  绕开组额度搬电就是把上面那个刷电漏洞放回来。resolveMembers 同理改为
  先查区块再取 BE, 解析不到就跳过而不是抛。

+++ src/main/java/com/miningdim/power/client/PowerCellScreen.java
@@ -9,7 +9,7 @@ import net.minecraft.world.entity.player.Inventory;
 
 import java.util.List;
 
-/** 三级储电共用界面：余额是主读数，进出功率各占一条细表。 */
+/** 三级储电共用界面：余额是主读数，进出功率各占一条细表。读数一律是整块所在储电组的聚合值。 */
 public final class PowerCellScreen extends AbstractPowerMachineScreen<PowerCellMenu> {
 
     private static final ResourceLocation BACKGROUND =
@@ -33,10 +33,10 @@ public final class PowerCellScreen extends AbstractPowerMachineScreen<PowerCellM
                                  int mouseX, int mouseY, float partialTick) {
         drawMeter(graphics, left + METER_X, top + MAIN_METER_Y, METER_WIDTH, MAIN_METER_HEIGHT,
                 menu.storedFe(), menu.capacityFe(), ENERGY_COLOR);
-        // 进出两条表以"本档最大传输速率"为满格基准, 靠单条表的填充比例就能看出电网是在充还是在放。
+        // 进出两条表以"整组最大传输速率"为满格基准, 靠单条表的填充比例就能看出电网是在充还是在放。
         // 基准必须是这个固定值而不是两条读数的较大者: 取较大者会让其中一条永远画成满格, 两条之间也就
-        // 只剩相对关系, 读不出"离本档上限还有多远"。
-        int flowScale = Math.max(1, menu.transferFePerTick());
+        // 只剩相对关系, 读不出"离本组上限还有多远"。
+        long flowScale = Math.max(1L, menu.transferFePerTick());
         drawMeter(graphics, left + METER_X, top + RECEIVED_METER_Y, METER_WIDTH, FLOW_METER_HEIGHT,
                 menu.lastReceivedFe(), flowScale, PROCESS_COLOR);
         drawMeter(graphics, left + METER_X, top + EXTRACTED_METER_Y, METER_WIDTH, FLOW_METER_HEIGHT,
